// osa9/patientor-backend/src/types.ts
export type Diagnose = {
    code: string;
    name: string;
    latin?: string;
};

export enum Gender {
    MALE = 'male',
    FEMALE = 'female',
    OTHER = 'other',
}

export type Patient = {
    id: string;
    name: string;
    dateOfBirth: string;
    ssn: string;
    gender: Gender;
    occupation: string;
};

export type NewPatientEntry = Omit<Patient, 'id'>;

export type DescreetPatient = Omit<Patient, 'ssn'>;